separate MCMC features and exclude MCMC from coverage and tests
