+++ setup.cfg
@@ -1,2 +1,5 @@
 [nosetests]
-exclude=test_(extractor_ctio_planetary_nebula|astrometry)
+exclude=test_(extractor_ctio_planetary_nebula|astrometry|mcmc)
+
+[coverage:run]
+omit=spectractor/fit/mcmc.py
