+++ spectractor/fit/fitter.py
@@ -1,15 +1,11 @@
 from iminuit import Minuit
 from scipy import optimize
-from schwimmbad import MPIPool
-import emcee
 import time
 import matplotlib.pyplot as plt
 from matplotlib.ticker import MaxNLocator
 import numpy as np
-import sys
 import os
 import json
-import multiprocessing
 from dataclasses import dataclass
 from typing import Optional, Union
 
@@ -17,7 +13,6 @@ from spectractor import parameters
 from spectractor.config import set_logger
 from spectractor.tools import (formatting_numbers, compute_correlation_matrix, plot_correlation_matrix_simple,
                                NumpyArrayEncoder)
-from spectractor.fit.statistics import Likelihood
 
 
 @dataclass()
@@ -1118,384 +1113,6 @@ class FitWorkspace:
         self.my_logger.info(f"\n\tSave gradient descent log {output_filename}.")
 
 
-class MCMCFitWorkspace(FitWorkspace):
-
-    def __init__(self, params, file_name="", nwalkers=18, nsteps=1000, burnin=100, nbins=10,
-                 verbose=False, plot=False, live_fit=False, truth=None):
-        """Generic class to create a fit workspace with parameters, bounds and general fitting methods.
-
-        Parameters
-        ----------
-        params: FitParameters
-            The parameters to fit to data.
-        file_name: str, optional
-            The generic file name to save results. If file_name=="", nothing is saved ond disk (default: "").
-        nwalkers: int, optional
-            Number of walkers for MCMC exploration (default: 18).
-        nsteps: int, optional
-            Number of steps for MCMC exploration (default: 1000).
-        burnin: int, optional
-            Number of burn-in steps for MCMC exploration (default: 100).
-        nbins: int, optional
-            Number of bins to make histograms after MCMC exploration (default: 10).
-        verbose: bool, optional
-            Level of verbosity (default: False).
-        plot: bool, optional
-            Level of plotting (default: False).
-        live_fit: bool, optional
-            If True, model, data and residuals plots are made along the fitting procedure (default: False).
-        truth: array_like, optional
-            Array of true parameters (default: None).
-
-        Examples
-        --------
-        >>> params = FitParameters(values=[1, 1, 1, 1, 1])
-        >>> w = MCMCFitWorkspace(params)
-        >>> w.nwalkers
-        18
-        """
-        FitWorkspace.__init__(self, params, file_name=file_name, verbose=verbose, plot=plot, live_fit=live_fit, truth=truth)
-        self.my_logger = set_logger(self.__class__.__name__)
-        self.nwalkers = max(2 * self.params.ndim, nwalkers)
-        self.nsteps = nsteps
-        self.nbins = nbins
-        self.burnin = burnin
-        self.start = []
-        self.likelihood = np.array([[]])
-        self.gelmans = np.array([])
-        self.chains = np.array([[]])
-        self.lnprobs = np.array([[]])
-        self.flat_chains = np.array([[]])
-        self.valid_chains = [False] * self.nwalkers
-        self.global_average = None
-        self.global_std = None
-        self.use_grid = False
-        if self.filename != "":
-            if "." in self.filename:
-                self.emcee_filename = os.path.splitext(self.filename)[0] + "_emcee.h5"
-            else:
-                self.my_logger.warning("\n\tFile name must have an extension.")
-        else:
-            self.emcee_filename = "emcee.h5"
-
-    def set_start(self, percent=0.02, a_random=1e-5):
-        """Set the random starting points for MCMC exploration.
-
-        A set of parameters are drawn with a uniform distribution between +/- percent times the starting guess.
-        For null guess parameters, starting points are drawn from a uniform distribution between +/- a_random.
-
-        Parameters
-        ----------
-        percent: float, optional
-            Percent of the guess parameters to set the uniform interval to draw random points (default: 0.02).
-        a_random: float, optional
-            Absolute value to set the +/- uniform interval to draw random points
-            for null guess parameters (default: 1e-5).
-
-        Returns
-        -------
-        start: np.array
-            Array of starting points of shape (ndim, nwalkers).
-
-        """
-        self.start = np.array([np.random.uniform(self.params.values[i] - percent * self.params.values[i],
-                                                 self.params.values[i] + percent * self.params.values[i],
-                                                 self.nwalkers) for i in range(self.params.ndim)]).T
-        self.start[self.start == 0] = a_random * np.random.uniform(-1, 1)
-        return self.start
-
-    def load_chains(self):
-        """Load the MCMC chains from a hdf5 file. The burn-in points are not rejected at this stage.
-
-        Returns
-        -------
-        chains: np.array
-            Array of the chains.
-        lnprobs: np.array
-            Array of the logarithmic posterior probability.
-
-        """
-        self.chains = [[]]
-        self.lnprobs = [[]]
-        self.nsteps = 0
-        # tau = -1
-        reader = emcee.backends.HDFBackend(self.emcee_filename)
-        try:
-            tau = reader.get_autocorr_time()
-        except emcee.autocorr.AutocorrError:
-            tau = -1
-        self.chains = reader.get_chain(discard=0, flat=False, thin=1)
-        self.lnprobs = reader.get_log_prob(discard=0, flat=False, thin=1)
-        self.nsteps = self.chains.shape[0]
-        self.nwalkers = self.chains.shape[1]
-        print(f"Auto-correlation time: {tau}")
-        print(f"Burn-in: {self.burnin}")
-        print(f"Chains shape: {self.chains.shape}")
-        print(f"Log prob shape: {self.lnprobs.shape}")
-        return self.chains, self.lnprobs
-
-    def build_flat_chains(self):
-        """Flatten the chains array and apply burn-in.
-
-        Returns
-        -------
-        flat_chains: np.array
-            Flat chains.
-
-        """
-        self.flat_chains = self.chains[self.burnin:, self.valid_chains, :].reshape((-1, self.params.ndim))
-        return self.flat_chains
-
-    def analyze_chains(self):
-        """Load the chains, build the probability densities for the parameters, compute the best fitting values
-        and the uncertainties and covariance matrices, and plot.
-
-        """
-        self.load_chains()
-        self.set_chain_validity()
-        self.convergence_tests()
-        self.build_flat_chains()
-        self.likelihood = self.chain2likelihood()
-        self.params.cov = self.likelihood.cov_matrix
-        self.params.values = self.likelihood.mean_vec
-        self.simulate(*self.params.values)
-        self.plot_fit()
-        figure_name = os.path.splitext(self.emcee_filename)[0] + '_triangle.pdf'
-        self.likelihood.triangle_plots(output_filename=figure_name)
-
-    def chain2likelihood(self, pdfonly=False, walker_index=-1):
-        """Convert the chains to a psoterior probability density function via histograms.
-
-        Parameters
-        ----------
-        pdfonly: bool, optional
-            If True, do not compute the covariances and the 2D correlation plots (default: False).
-        walker_index: int, optional
-            The walker index to plot. If -1, all walkers are selected (default: -1).
-
-        Returns
-        -------
-        likelihood: np.array
-            Posterior density function.
-
-        """
-        if walker_index >= 0:
-            chains = self.chains[self.burnin:, walker_index, :]
-        else:
-            chains = self.flat_chains
-        rangedim = range(chains.shape[1])
-        centers = []
-        for i in rangedim:
-            centers.append(np.linspace(np.min(chains[:, i]), np.max(chains[:, i]), self.nbins - 1))
-        likelihood = Likelihood(centers, labels=self.params.labels, axis_names=self.params.axis_names, truth=self.params.truth)
-        if walker_index < 0:
-            for i in rangedim:
-                likelihood.pdfs[i].fill_histogram(chains[:, i], weights=None)
-                if not pdfonly:
-                    for j in rangedim:
-                        if i != j:
-                            likelihood.contours[i][j].fill_histogram(chains[:, i], chains[:, j], weights=None)
-            output_file = ""
-            if self.filename != "":
-                output_file = os.path.splitext(self.filename)[0] + "_bestfit.txt"
-            likelihood.stats(output=output_file)
-        else:
-            for i in rangedim:
-                likelihood.pdfs[i].fill_histogram(chains[:, i], weights=None)
-        return likelihood
-
-    def compute_local_acceptance_rate(self, start_index, last_index, walker_index):
-        """Compute the local acceptance rate in a chain.
-
-        Parameters
-        ----------
-        start_index: int
-            Beginning index.
-        last_index: int
-            End index.
-        walker_index: int
-            Index of the walker.
-
-        Returns
-        -------
-        freq: float
-            The acceptance rate.
-
-        """
-        frequences = []
-        test = -2 * self.lnprobs[start_index, walker_index]
-        counts = 1
-        for index in range(start_index + 1, last_index):
-            chi2 = -2 * self.lnprobs[index, walker_index]
-            if np.isclose(chi2, test):
-                counts += 1
-            else:
-                frequences.append(float(counts))
-                counts = 1
-                test = chi2
-        frequences.append(counts)
-        return 1.0 / np.mean(frequences)
-
-    def set_chain_validity(self):
-        """Test the validity of a chain: reject chains whose chi2 is far from the mean of the others.
-
-        Returns
-        -------
-        valid_chains: list
-            List of boolean values, True if the chain is valid, or False if invalid.
-
-        """
-        nchains = [k for k in range(self.nwalkers)]
-        chisq_averages = []
-        chisq_std = []
-        for k in nchains:
-            chisqs = -2 * self.lnprobs[self.burnin:, k]
-            # if np.mean(chisqs) < 1e5:
-            chisq_averages.append(np.mean(chisqs))
-            chisq_std.append(np.std(chisqs))
-        self.global_average = np.mean(chisq_averages)
-        self.global_std = np.mean(chisq_std)
-        self.valid_chains = [False] * self.nwalkers
-        for k in nchains:
-            chisqs = -2 * self.lnprobs[self.burnin:, k]
-            chisq_average = np.mean(chisqs)
-            chisq_std = np.std(chisqs)
-            if 3 * self.global_std + self.global_average < chisq_average < 1e5:
-                self.valid_chains[k] = False
-            elif chisq_std < 0.1 * self.global_std:
-                self.valid_chains[k] = False
-            else:
-                self.valid_chains[k] = True
-        return self.valid_chains
-
-    def convergence_tests(self):
-        """Compute the convergence tests (Gelman-Rubin, acceptance rate).
-
-        """
-        chains = self.chains[self.burnin:, :, :]  # .reshape((-1, self.ndim))
-        nchains = [k for k in range(self.nwalkers)]
-        fig, ax = plt.subplots(self.params.ndim + 1, 2, figsize=(16, 7), sharex='all')
-        fontsize = 8
-        steps = np.arange(self.burnin, self.nsteps)
-        # Chi2 vs Index
-        print("Chisq statistics:")
-        for k in nchains:
-            chisqs = -2 * self.lnprobs[self.burnin:, k]
-            text = f"\tWalker {k:d}: {float(np.mean(chisqs)):.3f} +/- {float(np.std(chisqs)):.3f}"
-            if not self.valid_chains[k]:
-                text += " -> excluded"
-                ax[self.params.ndim, 0].plot(steps, chisqs, c='0.5', linestyle='--')
-            else:
-                ax[self.params.ndim, 0].plot(steps, chisqs)
-            print(text)
-        # global_average = np.mean(-2*self.lnprobs[self.valid_chains, self.burnin:])
-        # global_std = np.std(-2*self.lnprobs[self.valid_chains, self.burnin:])
-        ax[self.params.ndim, 0].set_ylim(
-            [self.global_average - 5 * self.global_std, self.global_average + 5 * self.global_std])
-        # Parameter vs Index
-        print("Computing Parameter vs Index plots...")
-        for i in range(self.params.ndim):
-            ax[i, 0].set_ylabel(self.params.axis_names[i], fontsize=fontsize)
-            for k in nchains:
-                if self.valid_chains[k]:
-                    ax[i, 0].plot(steps, chains[:, k, i])
-                else:
-                    ax[i, 0].plot(steps, chains[:, k, i], c='0.5', linestyle='--')
-                ax[i, 0].get_yaxis().set_label_coords(-0.05, 0.5)
-        ax[self.params.ndim, 0].set_ylabel(r'$\chi^2$', fontsize=fontsize)
-        ax[self.params.ndim, 0].set_xlabel('Steps', fontsize=fontsize)
-        ax[self.params.ndim, 0].get_yaxis().set_label_coords(-0.05, 0.5)
-        # Acceptance rate vs Index
-        print("Computing acceptance rate...")
-        min_len = self.nsteps
-        window = 100
-        if min_len > window:
-            for k in nchains:
-                ARs = []
-                indices = []
-                for pos in range(self.burnin + window, self.nsteps, window):
-                    ARs.append(self.compute_local_acceptance_rate(pos - window, pos, k))
-                    indices.append(pos)
-                if self.valid_chains[k]:
-                    ax[self.params.ndim, 1].plot(indices, ARs, label=f'Walker {k:d}')
-                else:
-                    ax[self.params.ndim, 1].plot(indices, ARs, label=f'Walker {k:d}', c='gray', linestyle='--')
-                ax[self.params.ndim, 1].set_xlabel('Steps', fontsize=fontsize)
-                ax[self.params.ndim, 1].set_ylabel('Aceptance rate', fontsize=fontsize)
-                # ax[self.dim + 1, 2].legend(loc='upper left', ncol=2, fontsize=10)
-        # Parameter PDFs by chain
-        print("Computing chain by chain PDFs...")
-        for k in nchains:
-            likelihood = self.chain2likelihood(pdfonly=True, walker_index=k)
-            likelihood.stats(pdfonly=True, verbose=False)
-            # for i in range(self.dim):
-            # ax[i, 1].plot(likelihood.pdfs[i].axe.axis, likelihood.pdfs[i].grid, lw=var.LINEWIDTH,
-            #               label=f'Walker {k:d}')
-            # ax[i, 1].set_xlabel(self.axis_names[i])
-            # ax[i, 1].set_ylabel('PDF')
-            # ax[i, 1].legend(loc='upper right', ncol=2, fontsize=10)
-        # Gelman-Rubin test.py
-        if len(nchains) > 1:
-            step = max(1, (self.nsteps - self.burnin) // 20)
-            self.gelmans = []
-            print(f'Gelman-Rubin tests (burnin={self.burnin:d}, step={step:d}, nsteps={self.nsteps:d}):')
-            for i in range(self.params.ndim):
-                Rs = []
-                lens = []
-                for pos in range(self.burnin + step, self.nsteps, step):
-                    chain_averages = []
-                    chain_variances = []
-                    global_average = np.mean(self.chains[self.burnin:pos, self.valid_chains, i])
-                    for k in nchains:
-                        if not self.valid_chains[k]:
-                            continue
-                        chain_averages.append(np.mean(self.chains[self.burnin:pos, k, i]))
-                        chain_variances.append(np.var(self.chains[self.burnin:pos, k, i], ddof=1))
-                    W = np.mean(chain_variances)
-                    B = 0
-                    for n in range(len(chain_averages)):
-                        B += (chain_averages[n] - global_average) ** 2
-                    B *= ((pos + 1) / (len(chain_averages) - 1))
-                    R = (W * pos / (pos + 1) + B / (pos + 1) * (len(chain_averages) + 1) / len(chain_averages)) / W
-                    Rs.append(R - 1)
-                    lens.append(pos)
-                print(f'\t{self.params.labels[i]}: R-1 = {Rs[-1]:.3f} (l = {lens[-1] - 1:d})')
-                self.gelmans.append(Rs[-1])
-                ax[i, 1].plot(lens, Rs, lw=1, label=self.params.axis_names[i])
-                ax[i, 1].axhline(0.03, c='k', linestyle='--')
-                ax[i, 1].set_xlabel('Walker length', fontsize=fontsize)
-                ax[i, 1].set_ylabel('$R-1$', fontsize=fontsize)
-                ax[i, 1].set_ylim(0, 0.6)
-                # ax[self.dim, 3].legend(loc='best', ncol=2, fontsize=10)
-        self.gelmans = np.array(self.gelmans)
-        fig.tight_layout()
-        plt.subplots_adjust(hspace=0)
-        if parameters.DISPLAY:  # pragma: no cover
-            plt.show()
-        if parameters.PdfPages:
-            parameters.PdfPages.savefig()
-        figure_name = self.emcee_filename.replace('.h5', '_convergence.pdf')
-        print(f'Save figure: {figure_name}')
-        fig.savefig(figure_name, dpi=100)
-
-    def print_settings(self):
-        """Print the main settings of the FitWorkspace.
-
-        """
-        print('************************************')
-        print(f"Input file: {self.filename}\nWalkers: {self.nwalkers}\t Steps: {self.nsteps}")
-        print(f"Output file: {self.emcee_filename}")
-        print('************************************')
-
-
-def lnprob(p):  # pragma: no cover
-    global fit_workspace
-    lp = fit_workspace.lnprior(p)
-    if not np.isfinite(lp):
-        return -1e20
-    return lp + fit_workspace.lnlike(p)
-
-
 def gradient_descent(fit_workspace, epsilon, niter=10, xtol=1e-3, ftol=1e-3, with_line_search=True):
     """
 
@@ -1897,38 +1514,6 @@ def run_minimisation_sigma_clipping(fit_workspace, method="newton", epsilon=None
         else:
             my_logger.info(f'\n\tNo outliers detected at first iteration: break the sigma clipping iterations.')
             break
-
-
-def run_emcee(mcmc_fit_workspace, ln=lnprob):
-    my_logger = set_logger(__name__)
-    mcmc_fit_workspace.print_settings()
-    nsamples = mcmc_fit_workspace.nsteps
-    p0 = mcmc_fit_workspace.set_start()
-    filename = mcmc_fit_workspace.emcee_filename
-    backend = emcee.backends.HDFBackend(filename)
-    try:  # pragma: no cover
-        pool = MPIPool()
-        if not pool.is_master():
-            pool.wait()
-            sys.exit(0)
-        sampler = emcee.EnsembleSampler(mcmc_fit_workspace.nwalkers, mcmc_fit_workspace.ndim, ln, args=(),
-                                        pool=pool, backend=backend)
-        my_logger.info(f"\n\tInitial size: {backend.iteration}")
-        if backend.iteration > 0:
-            p0 = backend.get_last_sample()
-        if nsamples - backend.iteration > 0:
-            sampler.run_mcmc(p0, nsteps=max(0, nsamples - backend.iteration), progress=True)
-        pool.close()
-    except ValueError:
-        sampler = emcee.EnsembleSampler(mcmc_fit_workspace.nwalkers, mcmc_fit_workspace.params.ndim, ln, args=(),
-                                        threads=multiprocessing.cpu_count(), backend=backend)
-        my_logger.info(f"\n\tInitial size: {backend.iteration}")
-        if backend.iteration > 0:
-            p0 = sampler.get_last_sample()
-        for _ in sampler.sample(p0, iterations=max(0, nsamples - backend.iteration), progress=True, store=True):
-            continue
-    mcmc_fit_workspace.chains = sampler.chain
-    mcmc_fit_workspace.lnprobs = sampler.lnprobability
 
 
 class RegFitWorkspace(FitWorkspace):
