+++ tests/test_fitter.py
@@ -3,21 +3,18 @@ mpl.use('Agg')  # must be run first! But therefore requires noqa E402 on all oth
 
 from numpy.testing import run_module_suite  # noqa: E402
 import numpy as np  # noqa: E402
-from spectractor.fit.fitter import (FitParameters, MCMCFitWorkspace, run_minimisation,  # noqa: E402
-                                    run_emcee, run_minimisation_sigma_clipping)  # noqa: E402
+from spectractor.fit.fitter import (FitParameters, FitWorkspace, run_minimisation,  # noqa: E402
+                                    run_minimisation_sigma_clipping)  # noqa: E402
 from spectractor.config import set_logger  # noqa: E402
 from spectractor import parameters  # noqa: E402
 
-import os  # noqa: E402
 
+class LineFitWorkspace(FitWorkspace):
 
-class LineFitWorkspace(MCMCFitWorkspace):
-
-    def __init__(self, x, y, yerr, file_name="", nwalkers=18, nsteps=1000, burnin=100, nbins=10,
-                 verbose=0, plot=False, live_fit=False, truth=None):
+    def __init__(self, x, y, yerr, file_name="", verbose=0, plot=False, live_fit=False, truth=None):
         params = FitParameters(np.array([1, 1]), labels=["a", "b"], axis_names=["$a$", "$b$"],
                                bounds=[[-100, 100], [-100, 100]], truth=truth, filename=file_name)
-        MCMCFitWorkspace.__init__(self, params, file_name, nwalkers, nsteps, burnin, nbins, verbose, plot, live_fit)
+        FitWorkspace.__init__(self, params, file_name, verbose, plot, live_fit)
         self.my_logger = set_logger(self.__class__.__name__)
         self.x = x
         self.data = y
@@ -45,7 +42,7 @@ def test_fitworkspace(seed=42):
 
     # Do the fits
     file_name = "./outputs/test_linefitworkspace.txt"
-    w = LineFitWorkspace(x, y, yerr, file_name, truth=truth, nwalkers=20, nsteps=5000, burnin=1000, nbins=20)
+    w = LineFitWorkspace(x, y, yerr, file_name, truth=truth)
     run_minimisation(w, method="minimize")
     assert np.all([np.abs(w.params.values[i] - truth[i]) / sigma < 1 for i in range(w.params.ndim)])
     w.p = np.array([1, 1])
@@ -61,22 +58,6 @@ def test_fitworkspace(seed=42):
     w.p = np.array([1, 1])
     run_minimisation(w, method="newton")
     assert np.all([np.abs(w.params.values[i] - truth[i]) / sigma < 1 for i in range(w.params.ndim)])
-
-    # def lnprob(p):
-    #     lp = w.lnprior(p)
-    #     if not np.isfinite(lp):
-    #         return -1e20
-    #     return lp + w.lnlike(p)
-    #
-    # run_emcee(w, ln=lnprob)
-    # w.analyze_chains()
-    #
-    # assert w.chains.shape == (5000, 20, 2)
-    # assert np.all(w.gelmans < 0.03)
-    # assert os.path.exists(file_name.replace(".txt", "_emcee.h5"))
-    # assert os.path.exists(file_name.replace(".txt", "_emcee_convergence.pdf"))
-    # assert os.path.exists(file_name.replace(".txt", "_emcee_triangle.pdf"))
-    # assert np.all([np.abs(w.params.values[i] - truth[i]) / np.sqrt(w.params.cov[i, i]) < 3 for i in range(w.params.ndim)])
 
 
 def test_minimisation_sigma_clipping(seed=42):
@@ -101,7 +82,7 @@ def test_minimisation_sigma_clipping(seed=42):
     # Do the fits
     sigma = 5
     clip_niter = 4
-    w = LineFitWorkspace(x, y, yerr, file_name="", truth=truth, nwalkers=20, nsteps=5000, burnin=1000, nbins=20)
+    w = LineFitWorkspace(x, y, yerr, file_name="", truth=truth)
     run_minimisation_sigma_clipping(w, method="minimize", sigma_clip=sigma, niter_clip=clip_niter)
     assert np.all([np.abs(w.params.values[i] - truth[i]) / sigma < 1 for i in range(w.params.ndim)])
     assert np.all(outliers == w.outliers)
